+++ modules/hanna-react/src/MainMenu/_Auxiliary.tsx
@@ -2,7 +2,8 @@ import React, { CSSProperties } from 'react';
 import { getIllustrationUrl } from '@reykjavik/hanna-utils/assets';
 
 import { Link } from '../_abstract/_Link';
-import type { MegaMenuPanel } from '../MainMenu';
+
+import { MegaMenuPanel } from './_PrimaryPanel';
 
 // ---------------------------------------------------------------------------